rename reserved cluster, move controller_utils function back
