+++ sky/data/storage_utils.py
@@ -1,12 +1,14 @@
 """Utility functions for the storage module."""
 import os
 import subprocess
-from typing import List
+from typing import Any, Dict, List
 
 import colorama
 
 from sky import exceptions
 from sky import sky_logging
+from sky.utils import log_utils
+from sky.utils.cli_utils import status_utils
 
 logger = sky_logging.init_logger(__name__)
 
@@ -15,6 +17,49 @@ _FILE_EXCLUSION_FROM_GITIGNORE_FAILURE_MSG = (
     'specified in .gitignore will be uploaded '
     'to the cloud storage for {path!r}'
     'due to the following error: {error_msg!r}')
+
+
+def format_storage_table(storages: List[Dict[str, Any]],
+                         show_all: bool = False) -> str:
+    """Format the storage table for display.
+
+    Args:
+        storage_table (dict): The storage table.
+
+    Returns:
+        str: The formatted storage table.
+    """
+    storage_table = log_utils.create_table([
+        'NAME',
+        'UPDATED',
+        'STORE',
+        'COMMAND',
+        'STATUS',
+    ])
+
+    for row in storages:
+        launched_at = row['launched_at']
+        if show_all:
+            command = row['last_use']
+        else:
+            command = status_utils.truncate_long_string(
+                row['last_use'], status_utils.COMMAND_TRUNC_LENGTH)
+        storage_table.add_row([
+            # NAME
+            row['name'],
+            # LAUNCHED
+            log_utils.readable_time_duration(launched_at),
+            # CLOUDS
+            ', '.join([s.value for s in row['store']]),
+            # COMMAND,
+            command,
+            # STATUS
+            row['status'].value,
+        ])
+    if storages:
+        return str(storage_table)
+    else:
+        return 'No existing storage.'
 
 
 def get_excluded_files_from_gitignore(src_dir_path: str) -> List[str]:
